+++ changes.md
@@ -1,3 +1,11 @@
+### 3.9.2+ (Build: 20201028)
+#### New plugins
+- Added `question/behavior/adaptivemultipart`
+- Added `question/type/formulas`
+
+#### Updates
+- Updated `mod/subcourse` to 9.0.1
+
 ### 3.9.2+ (Build: 20201016)
 #### Updates
 - Updated `theme/xtec2020` to 0.1.2